+++ controllers/operator/certs/certificates.go
@@ -183,9 +183,3 @@ func validateSelfManagedSSLCertsForStatefulSet(client kubernetesClient.Client, s
 
 	return workflow.OK()
 }
-
-// ToInternalClusterAuthName takes a hostname e.g. my-replica-set and converts
-// it into the name of the secret which will hold the internal clusterFile
-func ToInternalClusterAuthName(hostname string) string {
-	return fmt.Sprintf("%s-%s", hostname, util.ClusterFileName)
-}